added rgb2jpeg and inverse

+++ signals/lichtenstein.m
@@ -33,7 +33,7 @@ end;
 
 f=mfilename('fullpath');
 
-s=double(imread([f,'.png']));
+s=imread([f,'.png']);
